Add files via upload
The aim is to explore the strategies to win a game of connect 4 using a combi-
nation of a statistical and a strategic approach. Using the work of the original
solvers, I show and explain how powerful it is playing first and in the middle
column, in comparison to the other columns. I also show how player 2 can still
win with a knowledge and skill advantage